Move the user/group combo methods to a new file
This resolves a circular dependency between graph.py ->
service_account.py -> group.py -> graph.py... by making
service_account.py now rely on user_group.py instead of group.py
