+++ grouper/fe/handlers/permissions_revoke.py
@@ -1,8 +1,8 @@
 from grouper.fe.util import GrouperHandler
-from grouper.group import user_is_owner_of_group
 from grouper.models.audit_log import AuditLog
 from grouper.models.counter import Counter
 from grouper.models.permission_map import PermissionMap
+from grouper.user_group import user_is_owner_of_group
 from grouper.user_permissions import user_grantable_permissions
 from grouper.util import matches_glob
